requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
brotlipy==0.7.0
charset-normalizer==2.1.1
colorama==0.4.6
discord.py==2.1.0
frozenlist==1.3.3
multidict==6.0.4
Pillow==9.4.0
python-dotenv==0.21.0
yarl==1.8.2
requests~=2.28.1